Remove v4 signing and fix errors
This removes the v4 signing for Elasticsearch requests when running the
Fargate task. While the v4 signing should work in theory, we were never
able to get it function from within the Fargate task. We have updated
the Elasticsearch domain to allow requests from an IP range restricted
to the two subnets the task runs on.